+++ README.md
@@ -56,3 +56,54 @@
 Piece의 팩토리 메소드가 일반적으로 작성되어있었다면, (하나의 팩토리 메소드만이 외부에 대한 인터페이스였다면) 추상 메소드로 변경하기만 하면 되기 때문에 수정이 적었을 것이라고 생각했다. 팩토리 클래스를 따로 만드는 것. 수정에는 닫혀있고, 확장에는 열려있는 코드가 된다.
 확장성에 대한 고민 
 
+## temporal Todo
+- [x] Piece에 Position 필드 추가
+- [x] Board.move() 메소드 가독성 높이기
+- [x] game 클래스 분리
+- [ ] **view 재구현**
+- [ ] **main 재구현** - 컨트롤러 분리 -> game 클래스가 할 수 있을 것 같다.
+- [x] Game 클래스에서 Rank, Piece에 의존하지 않게 하기
+  - 점수 계산 로직 리팩토링 -> 성능 이슈 + Game에서 Rank
+- [x] Board, ChessGame 관련 테스트 코드 재작성
+- [x] 모든 하위 기물 테스트 코드 작성
+- [x] 모든 하위 기물 기능 구현
+- [x] Piece 관련 로직 모두 리팩토링
+- [x] isReachableDirection() 같은 메서드 뽑아내기가 가능할 지 생각해보기
+- [x] 예외 정의
+  1. Board 밖으로 이동
+  2. Blank 기물 이동
+  3. 아군 기물 존재하는 곳으로 이동
+  4. 해당 기물이 움직일 수 없는 곳으로 이동 시도
+
+   - [ ] Pawn 첫 이동 두 칸 구현 -> 필드 추가
+   - [x] 움직이려는 위치에 방해요소 여부 확인하기 -> board에서 구현하기
+   - [ ] king 잡으면 게임 종료 기능
+   - [ ] pawn 이동과 공격 로직 
+     - move 단에서, targetPosition이 상대 기물이면, capture direction에 있는지 확인
+     - 빈 곳이면, move direction에 있는지 확인 
+   - [ ] turn 구분 -> 현재 차례가 아닌 색의 기물을 움직이려 하면, ChessGame 단에서 예외 던지기
+   - [x] 점수 확인 command 구현
+   - [ ] 테스트코드 리팩토링
+     - parameterized test 적용하고, 테스트 이름 붙이기
+   - [ ] 패키지 구조 수정
+   - [ ] 이동 경로 상의 다른 기물이 있는 경우 예외 처리 기능
+   - [x] position string to Position -> App 단에서 모두 처리해서 주는게 맞지 않나?? 안쪽은 모두 좌표로만
+   - [ ] View를 분리한 후 Console에 찍어주는 역할의 view에서는 show하는 메서드로 바꾸기 -> 굳이?
+   - [ ] Piece의 position 필드 삭제하기... 숙원 사업
+
+## temporal 고민
+- 예외 처리를 어디에서 할 것인가? 지금 main의 move가 맡은 로직을 ChessGame으로 옮기고, main은 이것에 의존하고 던져진 예외에 따라 처리하는 역할을?
+- turn을 나타내는 변수를 Piece.Color로 해도 괜찮을까? 의미 상으로도 적절하고 코드도 간결해질 것 같긴 한데, Main이 Piece에 의존하는 것 같다. 라고 하기엔 단순히 내부 enum이라서 의존관계 때문에 문제가 생길 것 같지 않다.
+- 예외 종류가 적다.
+- 현재 Main의 멤버 변수들이 하는 역할을 ChessApp으로 분리했을 때의 장점은? Main은 entry point일 뿐이기 때문에 로직을 담지 않는 게 좋을까?
+  -> 로직을 담지 않는 게 맞는 것 같다. 앱 자체는 모듈화하고 실행하는 지점인 Main은 분리하는 것이~
+- 의존관계를 인터페이스로 추상화하기
+  -> 당장 생각할 수 있는 건, ChessApp의 인터페이스를 뽑아내서 웹에서 돌아갈 ChessApp으로의 확장을 생각하는 것
+  -> View 뽑아내기. 지금의 View의 역할은 콘솔에 찍어주는 것 뿐이다.
+  -> Exception 상속 구조 만들기
+- Piece를 연속적인 이동을 하는 기물과 한 칸만 이동하는 기물로 구분해서 상속하게 하는 것의 장점은? 
+  - 메소드 일반화 가능
+  - board, rank에서 이동 관련 로직 깔끔하게 검증하기 가능
+- position string to Position -> App 단에서 모두 처리해서 주는게 맞는 것 같다. 
+  입력된 string 값에 대해서는 App만이 알고 있고, 그 아래는 모두 Position 객체로만 교류한다. -> 무분별하게 ChessPositionParser에 의존하고, 중복 코드가 발생하는 문제가 해결된다.
+- 웹으로 구현되는 경우를 생각해서 view를 인터페이스로 뽑아내는 방안에 대해서 고민했다. 템플릿에 넘겨주는 view는 그 형태나 return type도 많이 다를 것 같아 인터페이스를 만들기만 하고 구현할 메소드를 제한하진 않았다.
