+++ src/main/java/softeer2nd/chess/domain/pieces/Blank.java
@@ -1,6 +1,6 @@
-package softeer2nd.chess.pieces;
+package softeer2nd.chess.domain.pieces;
 
-import softeer2nd.chess.Position;
+import softeer2nd.chess.domain.Position;
 import softeer2nd.chess.exceptions.BlankMoveException;
 
 public class Blank extends Piece {
