+++ src/test/java/softeer2nd/chess/ChessViewTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import softeer2nd.chess.exceptions.IllegalCommandException;
+import softeer2nd.chess.view.ChessView;
 
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
