+++ src/test/java/softeer2nd/chess/pieces/KingTest.java
@@ -3,10 +3,10 @@ package softeer2nd.chess.pieces;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import softeer2nd.chess.Position;
+import softeer2nd.chess.domain.Position;
+import softeer2nd.chess.domain.pieces.King;
 import softeer2nd.chess.exceptions.OutOfPieceRangeException;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 class KingTest {
